ConcordiaMaps/smart_planner/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
google-generativeai
google-genai